+++ celements-spring-mvc/src/main/java/com/celements/spring/security/IdentityServer.java
@@ -1,6 +1,10 @@
 package com.celements.spring.security;
 
 import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+
+import org.springframework.security.authentication.AuthenticationManager;
+import org.xwiki.model.reference.WikiReference;
 
 public interface IdentityServer {
 
@@ -9,5 +13,8 @@ public interface IdentityServer {
 
   @NotEmpty
   String getRealm();
+
+  @NotNull
+  AuthenticationManager getAuthenticationManagerForWiki(WikiReference wikiRef);
 
 }
